requirements.txt
amqp==5.3.1
anyio==4.10.0
argcomplete==3.6.2
asgiref==3.9.1
azure-core==1.36.0
azure-storage-blob==12.27.1
billiard==4.2.1
boto3==1.40.19
botocore==1.40.19
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
cfn-flip==1.3.0
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.6
dj-database-url==3.0.1
Django==4.2.23
django-cors-headers==4.9.0
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
durationpy==0.10
exceptiongroup==1.3.0
firebase_admin==7.1.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hjson==3.1.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
isodate==0.7.2
itypes==1.2.0
Jinja2==3.1.6
jmespath==1.0.1
kappa==0.6.0
kombu==5.5.4
libadvian==1.9.0
MarkupSafe==3.0.3
msgpack==1.1.1
openapi-codec==1.3.2
packaging==25.0
pillow==11.3.0
placebo==0.9.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
s3transfer==0.13.1
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==13.2.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
troposphere==4.9.4
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.37.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
